Oops... Forgot to include the vcproj with the changes for including input.cpp/.h >_<
Also, input is now using a virtual trackpad instead of a virtual joystick.  This, and most buttons for menus, are implemented in all menus.  Only not working is in-level.

Also includes some minor cleanup (removing all references to the problematic keysactive, as well as some unnecessary printf's), and some minor struct changes to make code-entry easier.